requirements.txt
Click==6.7
biopython==1.70
numpy==1.14.1
tqdm==4.7.2